+++ crates/stackable-webhook/CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+### Changed
+
+- Check for certificate rotation every 5 minutes to prevent expiry due to monotonic and wall clock divergence ([#1175]).
+
+[#1175]: https://github.com/stackabletech/operator-rs/pull/1175
+
 ## [0.9.0] - 2026-02-03
 
 ### Added